+++ include/vrv/vrvdef.h
@@ -146,6 +146,7 @@ enum ClassId {
     DIR,
     DYNAM,
     FERMATA,
+    GLISS,
     HAIRPIN,
     HARM,
     MORDENT,