+++ src/iomei.cpp
@@ -50,6 +50,7 @@
 #include "fig.h"
 #include "ftrem.h"
 #include "functorparams.h"
+#include "gliss.h"
 #include "gracegrp.h"
 #include "hairpin.h"
 #include "halfmrpt.h"
@@ -509,6 +510,10 @@ bool MEIOutput::WriteObject(Object *object)
     else if (object->Is(FTREM)) {
         m_currentNode = m_currentNode.append_child("fTrem");
         WriteFTrem(m_currentNode, dynamic_cast<FTrem *>(object));
+    }
+    else if (object->Is(GLISS)) {
+        m_currentNode = m_currentNode.append_child("gliss");
+        WriteGliss(m_currentNode, dynamic_cast<Gliss *>(object));
     }
     else if (object->Is(GRACEGRP)) {
         m_currentNode = m_currentNode.append_child("graceGrp");
@@ -1232,6 +1237,18 @@ void MEIOutput::WriteFermata(pugi::xml_node currentNode, Fermata *fermata)
     fermata->WriteColor(currentNode);
     fermata->WriteFermataVis(currentNode);
     fermata->WritePlacement(currentNode);
+}
+
+void MEIOutput::WriteGliss(pugi::xml_node currentNode, Gliss *gliss)
+{
+    assert(gliss);
+
+    WriteControlElement(currentNode, gliss);
+    WriteTimeSpanningInterface(currentNode, gliss);
+    gliss->WriteColor(currentNode);
+    gliss->WriteLineRend(currentNode);
+    gliss->WriteLineRendBase(currentNode);
+    gliss->WriteNNumberLike(currentNode);
 }
 
 void MEIOutput::WriteHairpin(pugi::xml_node currentNode, Hairpin *hairpin)
@@ -3902,6 +3919,9 @@ bool MEIInput::ReadMeasureChildren(Object *parent, pugi::xml_node parentNode)
         else if (std::string(current.name()) == "fermata") {
             success = ReadFermata(parent, current);
         }
+        else if (std::string(current.name()) == "gliss") {
+            success = ReadGliss(parent, current);
+        }
         else if (std::string(current.name()) == "hairpin") {
             success = ReadHairpin(parent, current);
         }
@@ -4069,6 +4089,22 @@ bool MEIInput::ReadFermata(Object *parent, pugi::xml_node fermata)
 
     parent->AddChild(vrvFermata);
     ReadUnsupportedAttr(fermata, vrvFermata);
+    return true;
+}
+
+bool MEIInput::ReadGliss(Object *parent, pugi::xml_node gliss)
+{
+    Gliss *vrvGliss = new Gliss();
+    ReadControlElement(gliss, vrvGliss);
+
+    ReadTimeSpanningInterface(gliss, vrvGliss);
+    vrvGliss->ReadColor(gliss);
+    vrvGliss->ReadLineRend(gliss);
+    vrvGliss->ReadLineRendBase(gliss);
+    vrvGliss->ReadNNumberLike(gliss);
+
+    parent->AddChild(vrvGliss);
+    ReadUnsupportedAttr(gliss, vrvGliss);
     return true;
 }